pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-wa-review-tool"
version = "0.1.0"
description = "AWS Well-Architected Review Tool"
requires-python = ">=3.12"
dependencies = [
    "boto3",
    "pydantic>=2.0",
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "hypothesis",
    "pytest",
    "pytest-cov",
    "moto[all]",
]
infra = [
    "aws-cdk-lib",
    "constructs",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["cli*", "core*", "checks*", "backend*"]
